refact: 최대한 norm 수정, 일부 함수 파일 분리, 헤더에 정리 #13
