+++ lexer/lexer.h
@@ -6,7 +6,7 @@
 /*   By: jaesjeon <jaesjeon@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/19 21:11:59 by jaesjeon          #+#    #+#             */
-/*   Updated: 2022/08/30 23:33:37 by jaesjeon         ###   ########.fr       */
+/*   Updated: 2022/08/31 16:11:29 by jaesjeon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,16 +35,24 @@ typedef struct s_file
 	char			*name;
 	int				type;
 	int				match_flag;
-} t_file;
+}	t_file;
 
-int		lexer(t_lx_token **token_head, char *full_line, char *envp[]);
-void	interpreter(t_lx_token *token);
-void	print_token_list(t_lx_token *token_list);
-t_file	*get_files_cur_pwd(const char *cur_pwd, int dir_flag);
-int 	is_matching_file(char *input, t_file  * const files);
+// debug_function.c
+void			print_token_list(t_lx_token *token_list);
+// lexer.c
+int				lexer(t_lx_token **token_head, char *full_line, char *envp[]);
+// interpreter.c
+void			interpreter(t_lx_token *token);
 unsigned char	find_interpret_symbol(char **token_str, \
 												unsigned char target);
-char	*make_chunk_by_symbol(char **token_str, \
+// interpreter_middleware.c
+void			interpret_middleware(t_lx_token *token, char *chunk, \
+											unsigned char symbol_type);
+// dirent_utils.c
+t_file			*get_files_cur_pwd(const char *cur_pwd, int dir_flag);
+// find_files.c
+int				is_matching_file(char *input, t_file *const files);
+// interpreter_make_chunk.c
+char			*make_chunk_by_symbol(char **token_str, \
 					char *str_startpoint, unsigned char *symbol_type);
-
 #endif
