+++ origin_str_utils.c
@@ -1,0 +1,88 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   origin_str_utils.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jaesjeon <jaesjeon@student.42seoul.kr>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/31 15:05:05 by jaesjeon          #+#    #+#             */
+/*   Updated: 2022/08/31 15:57:36 by jaesjeon         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+size_t	ft_strlen(const char *s)
+{
+	size_t	i;
+
+	if (!s)
+		return (0);
+	i = 0;
+	while (s[i])
+		i++;
+	return (i);
+}
+
+char	*ft_strcpy(const char *start, const char *end)
+{
+	char	*ret;
+	size_t	len;
+	int		idx;
+
+	len = end - start;
+	ret = (char *)malloc(len + 1);
+	if (ret == NULL)
+		exit(GENERAL_EXIT_CODE);
+	idx = 0;
+	while (idx < (int)len)
+		ret[idx++] = *start++;
+	ret[idx] = '\0';
+	return (ret);
+}
+
+char	*ft_strchr(const char *s, int c)
+{
+	if (!s)
+		return (NULL);
+	while (*s != (char)c)
+	{
+		if (*s == '\0')
+			return (NULL);
+		s++;
+	}
+	return ((char *)s);
+}
+
+size_t	ft_strlcat(char *dst, char const *src, size_t dstsize)
+{
+	size_t	dst_len;
+	size_t	src_len;
+
+	dst_len = ft_strlen(dst);
+	src_len = ft_strlen(src);
+	if (dst_len + 1 > dstsize)
+		return (src_len + dstsize);
+	if (dst_len + src_len + 1 < dstsize)
+		dstsize = src_len + dst_len + 1;
+	dst[dstsize - 1] = '\0';
+	while (dstsize-- - 1 > dst_len)
+		dst[dstsize - 1] = src[dstsize - dst_len - 1];
+	return (dst_len + src_len);
+}
+
+size_t	ft_strlcpy(char *dst, char const *src, size_t dstsize)
+{
+	size_t	len;
+
+	len = ft_strlen(src);
+	if (dstsize > 0)
+	{
+		if (len + 1 < dstsize)
+			dstsize = len + 1;
+		dst[dstsize - 1] = '\0';
+		while (dstsize-- - 1)
+			dst[dstsize - 1] = src[dstsize - 1];
+	}
+	return (len);
+}
